fix: rename api directory to controllers
